chore: break out dice tests into separate files
